+++ tests/integration/cases/call_deep_dotted_method/Nim_heterogeneous_strategy_record_call@v2.nim
@@ -1,0 +1,11 @@
+{.warning[UnusedImport]:off.}
+type ClientType = object
+type ApiType = object
+    client: ClientType
+type ObjType = object
+    api: ApiType
+template post(self: ClientType; args: varargs[untyped]) = discard
+var obj: ObjType
+obj.api.client.post("hello")
+obj.api.client.post(42)
+obj.api.client.post(true)