+++ tests/integration/cases/call_dotted_method/Nim_heterogeneous_strategy_record_call@v2.nim
@@ -1,0 +1,9 @@
+{.warning[UnusedImport]:off.}
+type ClientType = object
+type AppType = object
+    client: ClientType
+template fetch(self: ClientType; args: varargs[untyped]) = discard
+var app: AppType
+app.client.fetch("hello")
+app.client.fetch(42)
+app.client.fetch(true)