+++ tests/integration/cases/call_ref_args_trivial_register/Nim_heterogeneous_strategy_record_call@v2.nim
@@ -1,0 +1,14 @@
+{.warning[UnusedImport]:off.}
+template process(args: varargs[untyped]) = discard
+var my_int = 1
+var my_bool = true
+var my_float = 3.14
+var my_list = @[
+    1,
+    2,
+    3
+]
+process(my_int, 42)
+process(my_bool, 7)
+process(my_float, 9)
+process(my_list, 1)