+++ tests/integration/cases/call_scalar_args/Nim_heterogeneous_strategy_record_call@v2.nim
@@ -1,0 +1,5 @@
+{.warning[UnusedImport]:off.}
+template process(args: varargs[untyped]) = discard
+process("hello")
+process(42)
+process(true)