+++ tests/integration/cases/call_scalar_args_uniform_second_slot/Nim_heterogeneous_strategy_record_call@v2.nim
@@ -1,0 +1,5 @@
+{.warning[UnusedImport]:off.}
+template process(args: varargs[untyped]) = discard
+process("hello", "a")
+process(42, "b")
+process(true, "c")