+++ tests/integration/cases/dict_mixed_scalars/Nim_heterogeneous_strategy_record_combined@v2.nim
@@ -1,0 +1,11 @@
+type Record0 = object
+    a: int
+    b: string
+var my_data = Record0(
+    a: 1,
+    b: "x"
+)
+my_data = Record0(
+    a: 1,
+    b: "x"
+)