+++ tests/integration/cases/record_epoch_datetime_i32_overflow/Nim_record_epoch_i32_overflow@v2.nim
@@ -1,0 +1,7 @@
+type Record0 = object
+    withinI32: int
+    beyondI32: int
+var my_data = Record0(
+    withinI32: 1705320000,
+    beyondI32: 4085195400
+)