+++ tests/integration/cases/record_sequence/Nim_heterogeneous_strategy_record@v2.nim
@@ -1,0 +1,9 @@
+type Record0 = object
+    id: int
+    label: string
+    tags: seq[string]
+var my_data = @[
+    Record0(id: 1, label: "first", tags: newSeq[string]()),
+    Record0(id: 2, label: "second", tags: newSeq[string]()),
+    Record0(id: 3, label: "third", tags: newSeq[string]())
+]